Private Inference using Gorilla hosted endpoint on Replicate (#162)
This PR provides a guide for users to set up private inference for
Gorilla models on Replicate. It includes instructions for configuring a
Gorilla model with Cog (Replicate's open source containerization tool),
building and pushing the Docker image, and running inference
programmatically using Replicate's Python client library. The guide
ensures users can easily deploy and interact with Gorilla models in a
private and controlled environment.

This addition is designed to facilitate a smoother workflow for users
seeking to utilize Gorilla models on a privately hosted endpoint via
Replicate, offering an alternative to the public zanino.berkeley.edu
endpoint.

Resolves #83

---------

Co-authored-by: Raman Varma <ramanvarma@Ramans-MacBook-Pro.local>
Co-authored-by: Shishir Patil <30296397+ShishirPatil@users.noreply.github.com>