[Bug Fix] Restrict AST checker on Boolean Variable (#256)
This PR removes any AST checker answers that contain the string literals
"True" and "False", thereby restricting the conditionals of the AST
checker to strictly boolean value.

Example:

`[true, "true"] -> [true]`

This PR will stack on top of #252

Will this change the results of leaderboard? YES

https://github.com/ShishirPatil/gorilla/pull/253 will updated the
leaderboard with this fix.