BFCL April 9th Release (#338)
This PR is for the BFCL April 9th release:

1. Bug fix in the evaluation dataset. This involves modifying both
prompts and function docs.
2. Bug fix for possible answers.

The detailed breakdown is attached below. If you spot any issue with our
evaluation dataset and/or possible answers, please feel free to raise an
issue!

| Test Category | Prompt/Func Doc Correction Count | Possible Answer
Correction Count |

|---------------------|-----------------------------|-----------------------------|
| Simple              | 3                           | 16 |
| Parallel             | 1                           | 16|
| Multiple              | 1                         | 11 |
| Parallel Multiple   | 10                          | 43 |

This PR **DOES** change the leaderboard score. We will update the
leaderboard website shortly, in PR #341

---------

Co-authored-by: Charlie Cheng-Jie Ji <charliechengjieji@berkeley.edu>
Co-authored-by: Fanjia Yan <fanjiayan@berkeley.edu>

---------

Co-authored-by: Charlie Cheng-Jie Ji <charliechengjieji@berkeley.edu>