BFCL April 10th Release  (#339)
This PR is for the BFCL April 10th release:

1. Introduce REST API sanity check for the executable test category. It
ensures that all the API endpoints involved during the execution
evaluation process are working properly. If any of them are not behaving
as expected, the evaluation process will be stopped by default as the
result will be inaccurate. Users can choose to bypass this check by
setting the `--skip-api-sanity-check` flag.


This PR **DOES NOT** change the leaderboard score.

---------

Co-authored-by: CharlieJCJ <charliechengjieji@berkeley.edu>