Update raft.py with default `p` to match paper
Change `p` which dictates the fraction of dataset with golden documents in them (vs) no golden documents. 

So, p = 0.8 means, for 80% of the train data set, `A* = Q + D* + D1 .. Dn` and for 20% of the train data set `A* = Q + D1 .. Dn` where `D*` are/is the golden document with the answer `A*`.