Leaderboard Update, in sync with BFCL April 27th Release (#391)
As mentioned in #390, in this PR, we fix some inconsistency issues in
the cost and latency calculation for open-source models, which are now
all calculated when serving the model with
[vLLM](https://github.com/vllm-project/vllm) using 8 V100 GPUs.
$$\text{Cost} = \text{Latency per 1000 function call} * (\text{8xV100
azure-pay-as-you-go-price per hour / 3600})$$

We want to thank the community for pointing out this oversight. Thanks
[@abacaj](https://twitter.com/abacaj) and
[@Teknium1](https://twitter.com/Teknium1) for initially raising the
issue, and thanks [@natikgadzhi](https://twitter.com/natikgadzhi)
[@HamelHusain](https://twitter.com/HamelHusain)
[@nicoritschel](https://twitter.com/nicoritschel)
[@winglian](https://twitter.com/winglian)
[@olafgeibig](https://twitter.com/olafgeibig) and many others for
joining the conversation. We are listening to community feedback and
continuously improving our Berkeley Function Calling Leaderboard.
Discussions like
[this](https://twitter.com/abacaj/status/1784003306508980250) serve as
great examples. Let us know what you want us to include next!

This PR DOES change the leaderboard scores for `costs` and `latency`,
but not `accuracy`.

---------

Co-authored-by: Charlie Cheng-Jie Ji
[charliechengjieji@berkeley.edu](mailto:charliechengjieji@berkeley.edu)
Co-authored-by: Fanjia Yan
[fanjiayan@berkeley.edu](mailto:fanjiayan@berkeley.edu)