Improve User Experience for Wagon Wheel (#399)
This PR aims to improve the user experience when interacting with the
wagon wheel on the BFCL leaderboard. Due to the huge number of entries,
the wagon wheel is having a hard time displaying them all. This PR
addresses the issue for laptop users. This issue still exists on mobile,
and we will tackle it later.

This PR **DOES NOT** change the leaderboard score.