BFCL May 6th Release (Dataset Bug Fix) (#412)
This PR is for the BFCL May 6th release.

Thanks to Brian Zhang at OpenAI for helping raise some of the issues.

1. Bug fix in the evaluation dataset. This involves modifying both
prompts and function docs.
2. Bug fix for possible answers.

The detailed breakdown is attached below. If you spot any issue with our
evaluation dataset and/or possible answers, please feel free to raise an
issue/PR!

| Test Category | Prompt/Func Doc Correction Count | Possible Answer
Correction Count |

|---------------------|-----------------------------|-----------------------------|
| Simple              | 6                           | 7 |
| Parallel             | 0                           | 0|
| Multiple              | 0                         | 1 |
| Parallel Multiple   | 0                          | 1 |

This PR **DOES** change the leaderboard score. We will update the
leaderboard website shortly, in a different PR.

---------

Co-authored-by: Charlie Cheng-Jie Ji <charliechengjieji@berkeley.edu>