Leaderboard Update, in sync with PR#407 (better handle float value comparison) (#459)
This PR updates the score on the leaderboard to reflect the changes in
the BFCL evaluation logic from #407. We lose the type restriction in
Python to accommodate the auto-conversion from int to float and allow
the use of `int` values for Python parameters expecting `float` values.
The detailed reasoning can be found at PR #407.