Fixes For NexusHandler (#437)
Hello! 

I just had a chance to take a look at the handler code for one of the
open source models for Raven, and noticed a few things that could be
fixed. I have made some fixes here, including fixes for converting the
typescript Function Calling DSL into Python Function Calling DSL. This
PR is for fixing the prompting issues only.

I believe this logic can be used for other Prompt-based models as well.
I notice you ask the model to generate Python outputs but give JSON
inputs for prompt-based FC; this can be applied to standardize the
prompting side for these other models as well.

I believe there are further fixes necessary on the AST parsing of the
outputs that can further improve the performance of the model. Happy to
help contribute more there!

I ran the model with these changes, and these are the new results I
notice:

| Category        | REPORTED | UPDATED |
|-----------------|----------|---------|
| Simple AST      | 68.91    | 80.5    |
| Multiple AST    | 76       | 86      |
| Parallel AST    | 30.5     | 42      |
| Parallel Multiple | 43.5    | 49      |
| Simple Exec     | 64.12    | 95      |
| Multiple Exec   | 82       | 92      |
| Parallel Exec   | 46       | 74      |
| Parallel Multiple I | 55    | 62.5    |
| Relevance       | 2.08     | 57.5    |
| **AVG**         | **52.01222222** | **70.94444444** |

Here are the results JSONs for each of the above with the new runs:
1. Simple AST:
[simple_score.json](https://github.com/ShishirPatil/gorilla/files/15380677/simple_score.json)
2. Multiple AST: 

[multiple_function_score.json](https://github.com/ShishirPatil/gorilla/files/15380683/multiple_function_score.json)
3. Parallel AST: 

[parallel_function_score.json](https://github.com/ShishirPatil/gorilla/files/15380686/parallel_function_score.json)
4. Parallel Multiple AST: 

[parallel_multiple_function_score.json](https://github.com/ShishirPatil/gorilla/files/15380691/parallel_multiple_function_score.json)
5. Simple Exec: 

[executable_simple_score.json](https://github.com/ShishirPatil/gorilla/files/15380694/executable_simple_score.json)
6. Multiple Exec: 

[executable_multiple_function_score.json](https://github.com/ShishirPatil/gorilla/files/15380696/executable_multiple_function_score.json)
7. Parallel Exec: 

[executable_parallel_function_score.json](https://github.com/ShishirPatil/gorilla/files/15380699/executable_parallel_function_score.json)
8. Parallel Multiple Exec: 

[executable_parallel_multiple_function_score.json](https://github.com/ShishirPatil/gorilla/files/15380700/executable_parallel_multiple_function_score.json)
9. Relevancy Score: 

[relevance_score.json](https://github.com/ShishirPatil/gorilla/files/15380703/relevance_score.json)

The average result improves by around 19% with these fixes, and likely
can improve more with more fixes on the AST side to eliminate false
negatives.

Regarding the AST fixes, there are several category of errors. For one
of the errors, as an example, please consider this:
```python
"model_result_decoded": [{"calculate_resistance": {"length": 5, "area": 0.01}}, {"calculate_resistance": {"length": 5, "area": 0.01, "resistivity": "aluminum"}}], "possible_answer": {"calculate_resistance_1": {"length": [5], "area": [0.01], "resistivity": ["copper"]}, "calculate_resistance_2": {"length": [5], "area": [0.01], "resistivity": ["aluminum"]}}}
```
It appears that the model has generated `calculate_resistance(5, 0.01);
calculate_resistance(5, 0.01, "aluminium")` but the ground truth is
`calculate_resistance(5, 0.01, "copper"); calculate_resistance(5, 0.01,
"aluminium")`, and therefore it is marked as wrong. But for Pythonic
DSL, the default values need not be generated, as "copper" is specified
as the default for this argument: `"resistivity": {"type": "string",
"description": "Resistivity of the material (Default: 'copper'). Allowed
values: 'copper', 'aluminum'"}}`. This is emblematic of other AST issues
that might need to be fixed as well, and can move accuracy higher to
reflect the real performance.

These AST issues might need to be fixed in another PR.

Very excited for the amazing work you have done here -- it is truly a
very comprehensive leaderboard!

Thank you so much!

---------

Co-authored-by: venkats <venkats@nexusflow.ai>