[BFCL] Standardize Model Name Among handler_map and eval_runner_helper (#439)
This PR standardize the model name among `handler_map` and
`eval_runner_helper`. In the past, the `handler_map` used the model name
that contains `/` while the `eval_runner_helper` takes the model name
that has the `/` substituted with `_`. This caused a lot of unnecessary
confusion when trying to add new models, so this PR addresses it by
letting the `eval_runner_helper` adapt to the `handler_map` format.
This PR **DOES NOT** affect the leaderboard score.