[BFCL] Fix Double-Casting Issue in model_handler for Java and JS category.  (#516)
When inferencing for Java or JS category, all function parameters are
already in `string` type (refer to [this
part](https://github.com/ShishirPatil/gorilla/blob/main/berkeley-function-call-leaderboard/model_handler/utils.py#L334-L374)).
So the model output must be `string` type as well to be correct.

The `decode_ast` method in some of the `model_handler` contains the
following code snippet:
```
if language == "Python":
    pass
else:
    # all values of the json are casted to string for java and javascript
    for key in params:
        params[key] = str(params[key])
```
The line `params[key] = str(params[key])` should not exist because it is
giving the model a 'second chance' when its parameter type is wrong (by
casting it to a `string` when it is not). This double type-casting will
introduce false positives to the evaluation result. As I mentioned in
#477, this PR addresses this issue.

Model Family Affected:

- OpenAI (FC only)
- Claude (FC only)
- Mistral (FC only)
- Gemini
- Cohere (FC only)

This will affect the leaderboard score. We will update it soon, in a
different PR.

---------

Co-authored-by: Shishir Patil <30296397+ShishirPatil@users.noreply.github.com>