Fixed the warning message "Setting `pad_token_id` to `eos_token_id`:1… (#110)
When running the falcon model, after the user prompt, the model response
is prefixed with this warning.
```
The attention mask and the pad token id were not set. As a consequence, you may observe unexpected behavior. Please pass your input's `attention_mask` to obtain reliable results. Setting `pad_token_id` to `eos_token_id`:11 for open-end generation. Fix for this issue
```

So Provided a fix.