[BFCL] Patch Generation Script for Locally Hosted OSS model (#537)
This PR updates the `inference` method in the `oss_handler` to improve
the generation task for locally-hosted models. Previously, we used Ray
for multi-node single-GPU inference (pipeline parallel), which is
uncommon as most setups are single machines with multiple GPUs. This
approach also led to out-of-memory errors for some large models.

In this PR:
- Ray is removed and replaced with vllm's built-in single-node multi-GPU
inference method (tensor parallel).
- The use of the `@torch.inference_mode()` decorator has been removed.