[BFCL] Add gemini-1.5-pro-002, gemini-1.5-pro-002-FC, gemini-1.5-pro-001, gemini-1.5-pro-001-FC, gemini-1.5-flash-002, gemini-1.5-flash-002-FC, gemini-1.0-pro-002, gemini-1.0-pro-002-FC (#658)
This PR adds the following new models to the leaderboard:

  - `gemini-1.5-pro-002`
  - `gemini-1.5-pro-002-FC`
  - `gemini-1.5-pro-001`
  - `gemini-1.5-pro-001-FC`
  - `gemini-1.5-flash-002`
  - `gemini-1.5-flash-002-FC`
  - `gemini-1.0-pro-002`
  - `gemini-1.0-pro-002-FC`

Note: All the code for Gemini prompting series inference are already
written out in #644, but they were not thoroughly verified at that time
due to limited bandwidth. This PR has the code tested and thus we
'officially' add support for the prompting version of the Gemini models.