[BFCL] Multi Turn Dataset Fix (Function Doc) (#722)
For functions used in the multi turn categories, we use automated
scripts to deterministically extract information from the function doc
string and compile the corresponding function docs. This PR addresses a
few issues with that pipeline.

1. [Fix] The doc strings are not in a uniform format, making info
extraction hard; they have been standardized.
2. [Fix] The compilation script is bugging and would sometimes output
wrong function doc with missing information.
3. [Fix] We have also add the compilation script to the `utils` folder,
as it may be useful to the community as well.

Following #719 , this is also part of the effort to thoroughly bug fix
the multi turn categories. We will have more PR coming in the next few
days.

---------

Co-authored-by: Amitoj Singh <amitojsingh2022@gmail.com>
Co-authored-by: Charlie Cheng-Jie Ji
<55744150+CharlieJCJ@users.noreply.github.com>
Co-authored-by: Fanjia-Yan
<78303449+Fanjia-Yan@users.noreply.github.com>
Co-authored-by: VishnuSuresh27
<112032533+VishnuSuresh27@users.noreply.github.com>

---------

Co-authored-by: amitojsingh2022 <amitojsingh2022@gmail.com>