[BFCL] Multi Turn Dataset Fix (Base Category) (#723)
This PR fixes the ambiguous prompt issue and some wrong ground truth
issues for the multi_turn_base category. After this PR, the
multi_turn_base entries should be bug-free.

Following #719 and #722 , this is also part of the effort to thoroughly
bug fix the multi turn categories. We will have more PR coming in the
next few days.

---------

Co-authored-by: Charlie Cheng-Jie Ji
<55744150+CharlieJCJ@users.noreply.github.com>
Co-authored-by: Fanjia-Yan
<78303449+Fanjia-Yan@users.noreply.github.com>
Co-authored-by: VishnuSuresh27
<112032533+VishnuSuresh27@users.noreply.github.com>