[BFCL] Update Eval Metric for Multi Turn Irrelevance Scenarios (#725)
In the current metric, for the `multi_turn_miss_func` and
`multi_turn_miss_param` categories, the model is expected to output no
function calls when a turn is missing necessary information (either a
relevant function or parameter). This mirrors the standard for
irrelevance detection in single-turn scenarios. However, multi-turn
interactions introduce additional complexity.

For instance, if the user’s request is "go to the ABC folder and display
content of the XYZ file" but the `cd` function isn’t provided, the model
might reasonably attempt exploratory actions (like calling `pwd` or `ls`
to check its current location) before recognizing that it cannot
complete the task as requested. Ultimately, the model should recognize
that the user’s task is unachievable given the context.

To address this, we've updated the metric: a dummy function,
`flag_task_unachievable`, will now be provided for every multi-turn
entry. If the model determines that one or more tasks are unachievable,
it should explicitly invoke this function. During evaluation, any entry
where the model calls this function will be marked as correct for
irrelevance detection, even if other functions were called beforehand.

In addition, this PR addresses #664. The execution result for each turn
(for both the model and the ground truth) is also included as part of
the score output files to help with debugging.

Following #719, #722 and #723, this is also part of the effort to
thoroughly bug fix the multi turn categories. We will have more PR
coming in the next few days.