[BFCL] Leaderboard Update - 2024/12/29 (Checkpoint 0cea216), Part 2 (#867)
This PR uses the same checkpoint as PR #845 and serves as a follow-up to
that PR. Previously, we were unable to obtain model results for some
models (namely, the Gemini 2.0 Flash series) due to rate limit
constraints. Those results are now available, and their scores have been
added in this PR.

The scores for other models remain unchanged.