Restrict GitHub Actions Workflow to Run Only on Source Repository (#1089)
This PR modifies the GitHub Actions workflow to ensure it only runs on
the main (source) repository and not on forks. Forks of this repository
do not have the necessary permissions or secrets for publishing to PiPy,
and attempting to run the workflow on those forks results in guaranteed
failures. This helps keep the CI clean.