[BFCL] Add support for Arch-Agent (#1078)
This PR adds Arch-Agent-{1.5B, 3B, 7B, 32B} to the leaderboard. 

**Note**
Note that we use YaRN scaling to deploy the models, and all Arch-Agent
models are evaluated with a context length of 64K. In the current
codebase, this requires adding extra fields for YaRN scaling when
spinning up servers and manually overwrite `self.max_context_length` in
the `base_oss_handler.py`.

---------

Co-authored-by: Huanzhi Mao <huanzhimao@gmail.com>