[BFCL] Fix wrong date in ground truth for `live_simple_205-116-13` (#1175)
PR #963 adjusted the date in the question of live_simple_205-116-13 to
be an accurate day of the week for the date given. However, it did not
adjust the answer based on the new day of the week.

We ask the model "What's the weather forecast for Berkeley on the
upcoming Saturday? Today is Tuesday April 25th 2023". Tuesday is the
25th, Wednesday the 26th, Thursday the 27th, Friday the 28th, and
Saturday the 29th. So, the answer to this question needs to be
"2023-04-29". The previous answer of "2023-04-27" was incorrect after
the merging of #963 and thus was improperly counting this wrong for
models that came up with the correct answer of "2023-04-29".

---------

Co-authored-by: Huanzhi Mao <huanzhimao@gmail.com>