[BFCL] - Additional Dataset Fixes, Builds off Issues 1133 PR (#1206)
multi_turn_base_11: uses find to find all the files and not ls .. could
specific prompt to not use find or use the most basic function –
separate PR incoming soon

multi_turn_base_15: change last entries to last entry

multi_turn_base_21: Given the proper context, and the given APIs in the
initial config, I believe this is clear enough for a model that is
coherent to use as context and call the diff() argument.

multi_turn_base_23: handled already in PR, echo doesn’t have ability to
create new file (need touch)

multi_turn_base_24: add the word and then to remove confusion and have
the ordering match the ground truth

multi_turn_base_33: remove the part about jotting them down to avoid any
confusion and keep the ground truth as correct

multi_turn_base_34: added the word whole to avoid confusion on using the
last line or not, keeping the ground truth to be correct

multi_turn_base_39: added the start as Enter the project and populate…
this was it knows that cd is required and keeps everything consistent
with the possible answers

Multi_turn_base_93: I don’t agree here, I believe that the model is
wrong because it ignores the “but” argument in the first turn saying to
check the fuel first. The second turn it should after checking then fill
it up fully.

Multi_turn_base_104: I think the extra steps are unnecessary and the
point of our prompt is to check if the model can follow our specific
instructions/prompting. A possible change could be that we add
get_watchlist at the start of the turn and then add to watchlist.
pending review

multi_turn_base_117: Double check, I think make_transaction is wrong
because we don’t have account_id so the fund function is correct to use
in this case.

multi_turn_base_162: to avoid confusion between integer and float
changed 2940 to 2940.15 for all respective entries and ground truths.

Multi_turn_base_178: This is fine, the booking id is in the initial
config so the ground truth is correct and the prompt is correct too.

Multi_turn_base_185: just added business CLASS journey to take away any
and all confusions that may have resulted in a model to check the
pricing for economy class

multi_turn_base_187: booking_record is in the initial config and is
provided, the model should have enough information to get the correct
ground truth for this prompt

multi_turn_base_194: I believe that if the model made the call without
the token, and then redid with the token and got everything, that is
still incorrect. There was no reason for the model to call the retrieve
invoice function without the token since it has access to that specific
token.

---------

Co-authored-by: Huanzhi Mao <huanzhimao@gmail.com>