Add model Nanbeige and Fix some data bugs (#1257)
Hi, thanks for the great work on BFCL!

This PR includes the following changes:

1. **Add two Nanbeige models**
- **Nanbeige4-3B-Thinking-2511 (open-source)**: Add model handler for
our open-source model. The weights are available on Hugging Face:
https://huggingface.co/Nanbeige/Nanbeige4-3B-Thinking-2511
- **Nanbeige3.5-Pro-Thinking (closed-source)**: The API URL and API key
used for evaluation are specified in the model handler file included in
this PR, and the model can also be accessed through our portal:
https://www.nanbeige.com/portal

2. **Fix two minor data issues**
- Fixed two small inconsistencies in the data to improve correctness and
stability (details are in the commit messages).

All changes are backward compatible and should not affect existing
workflows, other than improving the correctness of the data and
expanding the list of supported models.

If there is any preferred format for adding new model results or
additional checks you’d like us to run, I’m happy to update this PR
accordingly.

---------

Co-authored-by: Huanzhi Mao <huanzhimao@gmail.com>